requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
slowapi==0.1.9

pytest==8.3.4
httpx==0.28.1
SQLAlchemy==2.0.36
strawberry-graphql==0.287.3
pytest-asyncio==0.25.2